server/requirements.txt
fastapi
uvicorn
requests
pydantic
pydantic-settings
python-dotenv
supabase
geopy
pandas
